refactor(desktop): split store and measurement modules
