+++ desktop/src/components/workspace/shell/screen/MainSidebarPageShell.tsx
@@ -8,8 +8,8 @@ import { resolveStandardWorkspaceChromeClasses } from "@/lib/domain/preferences/
 import {
   WORKSPACE_SIDEBAR_MAX_WIDTH,
   WORKSPACE_SIDEBAR_MIN_WIDTH,
-  useWorkspaceUiStore,
-} from "@/stores/preferences/workspace-ui-store";
+} from "@/lib/domain/preferences/workspace-ui-state";
+import { useWorkspaceUiStore } from "@/stores/preferences/workspace-ui-store";
 import { SidebarUpdatePill } from "@/components/workspace/shell/sidebar/SidebarUpdatePill";
 import { MainSidebar } from "@/components/workspace/shell/sidebar/MainSidebar";
