+++ desktop/src/stores/preferences/workspace-ui-store.ts
@@ -1,21 +1,7 @@
 import type { SetStateAction } from "react";
 import { create } from "zustand";
+import type { ManualChatGroup } from "@/lib/domain/workspaces/tabs/manual-groups";
 import {
-  clearHiddenChatSessionIds,
-  rememberHiddenChatSessionId,
-  uniqueIds,
-} from "@/lib/domain/workspaces/tabs/visibility";
-import {
-  deleteManualChatGroup,
-  removeSessionsFromManualChatGroups,
-  sanitizeManualChatGroupsByWorkspace,
-  updateManualChatGroup,
-  upsertManualChatGroup,
-  type ManualChatGroup,
-} from "@/lib/domain/workspaces/tabs/manual-groups";
-import {
-  DEFAULT_SIDEBAR_WORKSPACE_TYPES,
-  resolveSidebarWorkspaceTypes,
   toggleSidebarWorkspaceTypeSelection,
   type SidebarWorkspaceVariant,
 } from "@/lib/domain/workspaces/sidebar/sidebar";
@@ -24,28 +10,23 @@ import {
   DEFAULT_RIGHT_PANEL_DURABLE_STATE,
   DEFAULT_RIGHT_PANEL_MATERIALIZED_STATE,
   normalizeRightPanelDurableState,
-  normalizeRightPanelMaterializedState,
   reconcileRightPanelWorkspaceState,
   type RightPanelDurableState,
   type RightPanelMaterializedState,
   type RightPanelWorkspaceState,
 } from "@/lib/domain/workspaces/shell/right-panel";
-import { migrateLegacyRightPanelWorkspaceState } from "@/lib/domain/workspaces/shell/right-panel-migration";
 import type { PendingChatActivation } from "@/lib/domain/workspaces/tabs/shell-activation";
 import {
-  parseWorkspaceShellTabKey,
   type WorkspaceShellIntentKey,
   type WorkspaceShellTabKey,
 } from "@/lib/domain/workspaces/tabs/shell-tabs";
 import {
-  sanitizeWorkspaceShellTabKeys,
-} from "@/lib/domain/workspaces/tabs/shell-file-seed";
-import { sameStringArray } from "@/lib/domain/workspaces/selection/workspace-keyed-preferences";
-import { readPersistedValue, persistValue } from "@/lib/infra/persistence/preferences-persistence";
-import {
-  isDebugMeasurementEnabled,
-  recordMeasurementDiagnostic,
-} from "@/lib/infra/measurement/debug-measurement";
+  clampWorkspaceSidebarWidth,
+  WORKSPACE_UI_DEFAULTS,
+  type PersistedWorkspaceUiState,
+} from "@/lib/domain/preferences/workspace-ui-state";
+import { createWorkspaceUiChatTabActions } from "@/stores/preferences/workspace-ui-chat-tab-actions";
+import { createWorkspaceUiShellActions } from "@/stores/preferences/workspace-ui-shell-actions";
 
 export interface WorkspaceUiState {
   _hydrated: boolean;
@@ -74,6 +55,7 @@ export interface WorkspaceUiState {
   recentlyHiddenChatSessionIdsByWorkspace: Record<string, string[]>;
   collapsedChatGroupsByWorkspace: Record<string, string[]>;
   manualChatGroupsByWorkspace: Record<string, ManualChatGroup[]>;
+  hydrate: (state: PersistedWorkspaceUiState) => void;
   archiveWorkspace: (id: string) => void;
   archiveWorkspaces: (ids: string[]) => void;
   unarchiveWorkspace: (id: string) => void;
@@ -179,53 +161,6 @@ export interface WorkspaceUiState {
   clearWorkspaceChatTabState: (workspaceId: string) => void;
 }
 
-/**
- * We have not launched yet, so persisted workspace UI state is allowed to
- * reset across identity-model changes.
- * v1: reset false-unread state caused by wall-clock interaction timestamps.
- * v2: reset user-facing workspace-keyed state for logical-workspace cutover.
- * v3: reset archived workspace ids after removing repositories stopped
- * polluting the archive model.
- * v4: add workspace-scoped right-panel preferences.
- * v5: add unified right-panel header order hints.
- * v6: current mainline schema with legacy unified right-panel preferences.
- * v7: split right-panel durable/materialized state and persist shell tab maps.
- * v8: make activeEntryKey/headerOrder the only right-panel selection/order
- *     model, removing durable toolOrder and terminal active/order fields.
- * v9: drop transient projected chat session ids from persisted workspace UI
- *     state; materialized last-viewed session ids own restart restore.
- */
-const WORKSPACE_UI_MIGRATION_VERSION = 9;
-export const WORKSPACE_SIDEBAR_DEFAULT_WIDTH = 280;
-export const WORKSPACE_SIDEBAR_MIN_WIDTH = 220;
-export const WORKSPACE_SIDEBAR_MAX_WIDTH = 420;
-
-export interface PersistedWorkspaceUiState {
-  migrationVersion?: number;
-  archivedWorkspaceIds: string[];
-  hiddenRepoRootIds: string[];
-  collapsedRepoGroups: string[];
-  showArchived: boolean;
-  threadsCollapsed: boolean;
-  sidebarOpen: boolean;
-  sidebarWidth: number;
-  rightPanelDurableByWorkspace: Record<string, RightPanelDurableState>;
-  rightPanelMaterializedByWorkspace: Record<string, RightPanelMaterializedState>;
-  activeShellTabKeyByWorkspace: Record<string, WorkspaceShellIntentKey | null>;
-  shellTabOrderByWorkspace: Record<string, WorkspaceShellTabKey[]>;
-  workspaceTypes: SidebarWorkspaceVariant[];
-  lastViewedAt: Record<string, string>;
-  lastViewedSessionByWorkspace: Record<string, string>;
-  lastViewedSessionErrorAtBySession: Record<string, string>;
-  workspaceLastInteracted: Record<string, string>;
-  dismissedSetupFailures: Record<string, boolean>;
-  finishSuggestionDismissalsByWorkspaceId: Record<string, string>;
-  visibleChatSessionIdsByWorkspace: Record<string, string[]>;
-  recentlyHiddenChatSessionIdsByWorkspace: Record<string, string[]>;
-  collapsedChatGroupsByWorkspace: Record<string, string[]>;
-  manualChatGroupsByWorkspace: Record<string, ManualChatGroup[]>;
-}
-
 export interface ShellIntentResult {
   changed: boolean;
   previousIntent: WorkspaceShellIntentKey | null;
@@ -233,527 +168,10 @@ export interface ShellIntentResult {
   epoch: number;
 }
 
-const WORKSPACE_UI_KEY = "workspace_ui";
-export const WORKSPACE_UI_DEFAULTS: PersistedWorkspaceUiState = {
-  archivedWorkspaceIds: [],
-  hiddenRepoRootIds: [],
-  collapsedRepoGroups: [],
-  showArchived: false,
-  threadsCollapsed: false,
-  sidebarOpen: false,
-  sidebarWidth: WORKSPACE_SIDEBAR_DEFAULT_WIDTH,
-  rightPanelDurableByWorkspace: {},
-  rightPanelMaterializedByWorkspace: {},
-  activeShellTabKeyByWorkspace: {},
-  shellTabOrderByWorkspace: {},
-  workspaceTypes: DEFAULT_SIDEBAR_WORKSPACE_TYPES,
-  lastViewedAt: {},
-  lastViewedSessionByWorkspace: {},
-  lastViewedSessionErrorAtBySession: {},
-  workspaceLastInteracted: {},
-  dismissedSetupFailures: {},
-  finishSuggestionDismissalsByWorkspaceId: {},
-  visibleChatSessionIdsByWorkspace: {},
-  recentlyHiddenChatSessionIdsByWorkspace: {},
-  collapsedChatGroupsByWorkspace: {},
-  manualChatGroupsByWorkspace: {},
-};
-
-function clampSidebarWidth(width: number): number {
-  return Math.min(WORKSPACE_SIDEBAR_MAX_WIDTH, Math.max(WORKSPACE_SIDEBAR_MIN_WIDTH, width));
-}
-
 function resolveStateValue<T>(value: SetStateAction<T>, current: T): T {
   return typeof value === "function"
     ? (value as (previousValue: T) => T)(current)
     : value;
-}
-
-async function readAll(): Promise<{ state: PersistedWorkspaceUiState; didMigrate: boolean }> {
-  let state: PersistedWorkspaceUiState;
-
-  const persisted = await readPersistedValue<PersistedWorkspaceUiState>(WORKSPACE_UI_KEY);
-  if (persisted) {
-    state = {
-      ...WORKSPACE_UI_DEFAULTS,
-      ...persisted,
-    };
-  } else {
-    state = {
-      archivedWorkspaceIds:
-        (await readPersistedValue<string[]>("archivedWorkspaceIds"))
-        ?? WORKSPACE_UI_DEFAULTS.archivedWorkspaceIds,
-      hiddenRepoRootIds: WORKSPACE_UI_DEFAULTS.hiddenRepoRootIds,
-      sidebarOpen: WORKSPACE_UI_DEFAULTS.sidebarOpen,
-      sidebarWidth: WORKSPACE_UI_DEFAULTS.sidebarWidth,
-      rightPanelDurableByWorkspace: WORKSPACE_UI_DEFAULTS.rightPanelDurableByWorkspace,
-      rightPanelMaterializedByWorkspace: WORKSPACE_UI_DEFAULTS.rightPanelMaterializedByWorkspace,
-      activeShellTabKeyByWorkspace: WORKSPACE_UI_DEFAULTS.activeShellTabKeyByWorkspace,
-      shellTabOrderByWorkspace: WORKSPACE_UI_DEFAULTS.shellTabOrderByWorkspace,
-      workspaceTypes: WORKSPACE_UI_DEFAULTS.workspaceTypes,
-      lastViewedAt:
-        (await readPersistedValue<Record<string, string>>("lastViewedAt"))
-        ?? WORKSPACE_UI_DEFAULTS.lastViewedAt,
-      lastViewedSessionByWorkspace:
-        (await readPersistedValue<Record<string, string>>("lastViewedSessionByWorkspace"))
-        ?? WORKSPACE_UI_DEFAULTS.lastViewedSessionByWorkspace,
-      lastViewedSessionErrorAtBySession:
-        WORKSPACE_UI_DEFAULTS.lastViewedSessionErrorAtBySession,
-      workspaceLastInteracted:
-        (await readPersistedValue<Record<string, string>>("workspaceLastInteracted"))
-        ?? WORKSPACE_UI_DEFAULTS.workspaceLastInteracted,
-      collapsedRepoGroups: WORKSPACE_UI_DEFAULTS.collapsedRepoGroups,
-      showArchived: WORKSPACE_UI_DEFAULTS.showArchived,
-      threadsCollapsed: WORKSPACE_UI_DEFAULTS.threadsCollapsed,
-      dismissedSetupFailures: WORKSPACE_UI_DEFAULTS.dismissedSetupFailures,
-      finishSuggestionDismissalsByWorkspaceId:
-        WORKSPACE_UI_DEFAULTS.finishSuggestionDismissalsByWorkspaceId,
-      visibleChatSessionIdsByWorkspace: WORKSPACE_UI_DEFAULTS.visibleChatSessionIdsByWorkspace,
-      recentlyHiddenChatSessionIdsByWorkspace:
-        WORKSPACE_UI_DEFAULTS.recentlyHiddenChatSessionIdsByWorkspace,
-      collapsedChatGroupsByWorkspace: WORKSPACE_UI_DEFAULTS.collapsedChatGroupsByWorkspace,
-      manualChatGroupsByWorkspace: WORKSPACE_UI_DEFAULTS.manualChatGroupsByWorkspace,
-    };
-  }
-
-  return migrateWorkspaceUiState(state);
-}
-
-export function migrateWorkspaceUiState(
-  input: PersistedWorkspaceUiState,
-): { state: PersistedWorkspaceUiState; didMigrate: boolean } {
-  const legacyInput = input as PersistedWorkspaceUiState & {
-    rightPanelByWorkspace?: Record<string, unknown>;
-    rightPanelWidthByWorkspace?: Record<string, number>;
-  };
-  const state = {
-    ...WORKSPACE_UI_DEFAULTS,
-    ...input,
-  };
-  let didMigrate = false;
-  const previousMigrationVersion = state.migrationVersion ?? 0;
-  if (previousMigrationVersion < 7) {
-    const migratedRightPanel = migrateLegacyRightPanelPreferences({
-      rightPanelByWorkspace: legacyInput.rightPanelByWorkspace,
-      rightPanelWidthByWorkspace: legacyInput.rightPanelWidthByWorkspace,
-    });
-    state.rightPanelDurableByWorkspace = {
-      ...migratedRightPanel.durableByWorkspace,
-      ...state.rightPanelDurableByWorkspace,
-    };
-    state.rightPanelMaterializedByWorkspace = {
-      ...migratedRightPanel.materializedByWorkspace,
-      ...state.rightPanelMaterializedByWorkspace,
-    };
-    didMigrate = true;
-  }
-  if (previousMigrationVersion < 3) {
-    state.archivedWorkspaceIds = [];
-    didMigrate = true;
-  }
-  if (previousMigrationVersion < 2) {
-    state.lastViewedAt = {};
-    state.lastViewedSessionByWorkspace = {};
-    state.workspaceLastInteracted = {};
-    didMigrate = true;
-  }
-  if (previousMigrationVersion < WORKSPACE_UI_MIGRATION_VERSION) {
-    state.migrationVersion = WORKSPACE_UI_MIGRATION_VERSION;
-    didMigrate = true;
-  }
-
-  // Migrate collapsedRepoGroups from Record<string, boolean> to string[]
-  if (!Array.isArray(state.collapsedRepoGroups)) {
-    const legacy = state.collapsedRepoGroups as unknown as Record<string, boolean>;
-    state.collapsedRepoGroups = Object.keys(legacy).filter((k) => legacy[k]);
-    didMigrate = true;
-  }
-
-  if (typeof state.sidebarOpen !== "boolean") {
-    state.sidebarOpen = WORKSPACE_UI_DEFAULTS.sidebarOpen;
-    didMigrate = true;
-  }
-
-  if (typeof state.showArchived !== "boolean") {
-    state.showArchived = WORKSPACE_UI_DEFAULTS.showArchived;
-    didMigrate = true;
-  }
-
-  if (typeof state.sidebarWidth !== "number" || Number.isNaN(state.sidebarWidth)) {
-    state.sidebarWidth = WORKSPACE_UI_DEFAULTS.sidebarWidth;
-    didMigrate = true;
-  }
-
-  const sanitizedRightPanelDurable = sanitizeRightPanelDurableByWorkspace(
-    state.rightPanelDurableByWorkspace,
-  );
-  if (JSON.stringify(sanitizedRightPanelDurable) !== JSON.stringify(state.rightPanelDurableByWorkspace)) {
-    state.rightPanelDurableByWorkspace = sanitizedRightPanelDurable;
-    didMigrate = true;
-  }
-
-  const sanitizedRightPanelMaterialized = sanitizeRightPanelMaterializedByWorkspace(
-    state.rightPanelMaterializedByWorkspace,
-  );
-  if (
-    JSON.stringify(sanitizedRightPanelMaterialized)
-    !== JSON.stringify(state.rightPanelMaterializedByWorkspace)
-  ) {
-    state.rightPanelMaterializedByWorkspace = sanitizedRightPanelMaterialized;
-    didMigrate = true;
-  }
-
-  const sanitizedActiveShellTabs = sanitizeActiveShellTabKeysByWorkspace(
-    state.activeShellTabKeyByWorkspace,
-  );
-  if (JSON.stringify(sanitizedActiveShellTabs) !== JSON.stringify(state.activeShellTabKeyByWorkspace)) {
-    state.activeShellTabKeyByWorkspace = sanitizedActiveShellTabs;
-    didMigrate = true;
-  }
-
-  const sanitizedShellOrder = sanitizeShellTabOrderByWorkspace(
-    state.shellTabOrderByWorkspace,
-  );
-  if (JSON.stringify(sanitizedShellOrder) !== JSON.stringify(state.shellTabOrderByWorkspace)) {
-    state.shellTabOrderByWorkspace = sanitizedShellOrder;
-    didMigrate = true;
-  }
-
-  const sanitizedVisibleChatSessions = sanitizeSessionIdArrayRecord(
-    state.visibleChatSessionIdsByWorkspace,
-  );
-  if (
-    JSON.stringify(sanitizedVisibleChatSessions)
-    !== JSON.stringify(state.visibleChatSessionIdsByWorkspace)
-  ) {
-    state.visibleChatSessionIdsByWorkspace = sanitizedVisibleChatSessions;
-    didMigrate = true;
-  }
-  if (!isStringArrayRecord(state.visibleChatSessionIdsByWorkspace)) {
-    state.visibleChatSessionIdsByWorkspace = WORKSPACE_UI_DEFAULTS.visibleChatSessionIdsByWorkspace;
-    didMigrate = true;
-  }
-
-  if (!isStringRecord(state.lastViewedSessionErrorAtBySession)) {
-    state.lastViewedSessionErrorAtBySession =
-      WORKSPACE_UI_DEFAULTS.lastViewedSessionErrorAtBySession;
-    didMigrate = true;
-  }
-
-  if (!isStringRecord(state.finishSuggestionDismissalsByWorkspaceId)) {
-    state.finishSuggestionDismissalsByWorkspaceId =
-      WORKSPACE_UI_DEFAULTS.finishSuggestionDismissalsByWorkspaceId;
-    didMigrate = true;
-  }
-
-  const sanitizedRecentlyHiddenChatSessions = sanitizeSessionIdArrayRecord(
-    state.recentlyHiddenChatSessionIdsByWorkspace,
-  );
-  if (
-    JSON.stringify(sanitizedRecentlyHiddenChatSessions)
-    !== JSON.stringify(state.recentlyHiddenChatSessionIdsByWorkspace)
-  ) {
-    state.recentlyHiddenChatSessionIdsByWorkspace = sanitizedRecentlyHiddenChatSessions;
-    didMigrate = true;
-  }
-  if (!isStringArrayRecord(state.recentlyHiddenChatSessionIdsByWorkspace)) {
-    state.recentlyHiddenChatSessionIdsByWorkspace =
-      WORKSPACE_UI_DEFAULTS.recentlyHiddenChatSessionIdsByWorkspace;
-    didMigrate = true;
-  }
-
-  const sanitizedCollapsedChatGroups = sanitizeSessionIdArrayRecord(
-    state.collapsedChatGroupsByWorkspace,
-  );
-  if (
-    JSON.stringify(sanitizedCollapsedChatGroups)
-    !== JSON.stringify(state.collapsedChatGroupsByWorkspace)
-  ) {
-    state.collapsedChatGroupsByWorkspace = sanitizedCollapsedChatGroups;
-    didMigrate = true;
-  }
-  if (!isStringArrayRecord(state.collapsedChatGroupsByWorkspace)) {
-    state.collapsedChatGroupsByWorkspace = WORKSPACE_UI_DEFAULTS.collapsedChatGroupsByWorkspace;
-    didMigrate = true;
-  }
-
-  const sanitizedManualGroups = sanitizeManualChatGroupsWithoutTransientSessions(
-    sanitizeManualChatGroupsByWorkspace(state.manualChatGroupsByWorkspace),
-  );
-  if (JSON.stringify(sanitizedManualGroups) !== JSON.stringify(state.manualChatGroupsByWorkspace)) {
-    state.manualChatGroupsByWorkspace = sanitizedManualGroups;
-    didMigrate = true;
-  }
-
-  const sanitizedLastViewedSessions = sanitizeLastViewedSessionByWorkspace(
-    state.lastViewedSessionByWorkspace,
-  );
-  if (
-    JSON.stringify(sanitizedLastViewedSessions)
-    !== JSON.stringify(state.lastViewedSessionByWorkspace)
-  ) {
-    state.lastViewedSessionByWorkspace = sanitizedLastViewedSessions;
-    didMigrate = true;
-  }
-
-  const clampedSidebarWidth = clampSidebarWidth(state.sidebarWidth);
-  if (clampedSidebarWidth !== state.sidebarWidth) {
-    state.sidebarWidth = clampedSidebarWidth;
-    didMigrate = true;
-  }
-
-  const resolvedWorkspaceTypes = resolveSidebarWorkspaceTypes(state.workspaceTypes);
-  if (
-    resolvedWorkspaceTypes.length !== state.workspaceTypes.length
-    || resolvedWorkspaceTypes.some((type, index) => type !== state.workspaceTypes[index])
-  ) {
-    state.workspaceTypes = resolvedWorkspaceTypes;
-    didMigrate = true;
-  }
-
-  return { state, didMigrate };
-}
-
-function selectPersistedSlice(state: WorkspaceUiState): PersistedWorkspaceUiState {
-  return {
-    migrationVersion: WORKSPACE_UI_MIGRATION_VERSION,
-    archivedWorkspaceIds: state.archivedWorkspaceIds,
-    hiddenRepoRootIds: state.hiddenRepoRootIds,
-    collapsedRepoGroups: state.collapsedRepoGroups,
-    showArchived: state.showArchived,
-    threadsCollapsed: state.threadsCollapsed,
-    sidebarOpen: state.sidebarOpen,
-    sidebarWidth: state.sidebarWidth,
-    rightPanelDurableByWorkspace: state.rightPanelDurableByWorkspace,
-    rightPanelMaterializedByWorkspace: state.rightPanelMaterializedByWorkspace,
-    activeShellTabKeyByWorkspace: sanitizeActiveShellTabKeysByWorkspace(
-      state.activeShellTabKeyByWorkspace,
-    ),
-    shellTabOrderByWorkspace: sanitizeShellTabOrderByWorkspace(
-      state.shellTabOrderByWorkspace,
-    ),
-    workspaceTypes: state.workspaceTypes,
-    lastViewedAt: state.lastViewedAt,
-    lastViewedSessionByWorkspace: sanitizeLastViewedSessionByWorkspace(
-      state.lastViewedSessionByWorkspace,
-    ),
-    lastViewedSessionErrorAtBySession: state.lastViewedSessionErrorAtBySession,
-    workspaceLastInteracted: state.workspaceLastInteracted,
-    dismissedSetupFailures: state.dismissedSetupFailures,
-    finishSuggestionDismissalsByWorkspaceId: state.finishSuggestionDismissalsByWorkspaceId,
-    visibleChatSessionIdsByWorkspace: sanitizeSessionIdArrayRecord(
-      state.visibleChatSessionIdsByWorkspace,
-    ),
-    recentlyHiddenChatSessionIdsByWorkspace: sanitizeSessionIdArrayRecord(
-      state.recentlyHiddenChatSessionIdsByWorkspace,
-    ),
-    collapsedChatGroupsByWorkspace: sanitizeSessionIdArrayRecord(
-      state.collapsedChatGroupsByWorkspace,
-    ),
-    manualChatGroupsByWorkspace: sanitizeManualChatGroupsWithoutTransientSessions(
-      sanitizeManualChatGroupsByWorkspace(state.manualChatGroupsByWorkspace),
-    ),
-  };
-}
-
-function isStringArrayRecord(value: unknown): value is Record<string, string[]> {
-  return typeof value === "object"
-    && value !== null
-    && Object.values(value).every((entry) =>
-      Array.isArray(entry) && entry.every((item) => typeof item === "string")
-    );
-}
-
-function isStringRecord(value: unknown): value is Record<string, string> {
-  return typeof value === "object"
-    && value !== null
-    && Object.values(value).every((entry) => typeof entry === "string");
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null;
-}
-
-function migrateLegacyRightPanelPreferences(args: {
-  rightPanelByWorkspace?: Record<string, unknown>;
-  rightPanelWidthByWorkspace?: Record<string, number>;
-}): {
-  durableByWorkspace: Record<string, RightPanelDurableState>;
-  materializedByWorkspace: Record<string, RightPanelMaterializedState>;
-} {
-  const legacyPanels = isRecord(args.rightPanelByWorkspace) ? args.rightPanelByWorkspace : {};
-  const legacyWidths = sanitizeRightPanelWidths(args.rightPanelWidthByWorkspace);
-  const workspaceIds = new Set([
-    ...Object.keys(legacyPanels),
-    ...Object.keys(legacyWidths),
-  ]);
-  const durableByWorkspace: Record<string, RightPanelDurableState> = {};
-  const materializedByWorkspace: Record<string, RightPanelMaterializedState> = {};
-
-  for (const workspaceId of workspaceIds) {
-    const legacyState = legacyPanels[workspaceId];
-    if (!isRecord(legacyState) && legacyWidths[workspaceId] === undefined) {
-      continue;
-    }
-    const { durableState, materializedState } = migrateLegacyRightPanelWorkspaceState({
-      state: legacyState,
-      width: legacyWidths[workspaceId],
-      isCloudWorkspaceSelected: true,
-    });
-    durableByWorkspace[workspaceId] = durableState;
-    materializedByWorkspace[workspaceId] = materializedState;
-  }
-
-  return { durableByWorkspace, materializedByWorkspace };
-}
-
-function sanitizeRightPanelDurableByWorkspace(
-  value: unknown,
-): Record<string, RightPanelDurableState> {
-  if (typeof value !== "object" || value === null) {
-    return {};
-  }
-
-  const next: Record<string, RightPanelDurableState> = {};
-  for (const [workspaceId, rawState] of Object.entries(value)) {
-    if (typeof rawState !== "object" || rawState === null) {
-      continue;
-    }
-    next[workspaceId] = normalizeRightPanelDurableState(
-      rawState as Partial<RightPanelDurableState>,
-    );
-  }
-  return next;
-}
-
-function sanitizeRightPanelMaterializedByWorkspace(
-  value: unknown,
-): Record<string, RightPanelMaterializedState> {
-  if (typeof value !== "object" || value === null) {
-    return {};
-  }
-
-  const next: Record<string, RightPanelMaterializedState> = {};
-  for (const [workspaceId, rawState] of Object.entries(value)) {
-    if (typeof rawState !== "object" || rawState === null) {
-      continue;
-    }
-    next[workspaceId] = normalizeRightPanelMaterializedState(
-      rawState as Partial<RightPanelMaterializedState>,
-      { isCloudWorkspaceSelected: true },
-    );
-  }
-  return next;
-}
-
-function sanitizeActiveShellTabKeysByWorkspace(
-  value: unknown,
-): Record<string, WorkspaceShellIntentKey | null> {
-  if (typeof value !== "object" || value === null) {
-    return {};
-  }
-  const next: Record<string, WorkspaceShellIntentKey | null> = {};
-  for (const [workspaceId, key] of Object.entries(value)) {
-    if (key === null) {
-      next[workspaceId] = null;
-      continue;
-    }
-    if (key === "chat-shell") {
-      next[workspaceId] = key;
-      continue;
-    }
-    if (
-      typeof key === "string"
-      && sanitizeWorkspaceShellTabKeys([key]).length === 1
-      && !isTransientChatTabKey(key)
-    ) {
-      next[workspaceId] = key;
-    }
-  }
-  return next;
-}
-
-function sanitizeShellTabOrderByWorkspace(
-  value: unknown,
-): Record<string, WorkspaceShellTabKey[]> {
-  if (typeof value !== "object" || value === null) {
-    return {};
-  }
-  const next: Record<string, WorkspaceShellTabKey[]> = {};
-  for (const [workspaceId, order] of Object.entries(value)) {
-    if (!Array.isArray(order)) {
-      continue;
-    }
-    const sanitized = sanitizeWorkspaceShellTabKeys(order)
-      .filter((key) => !isTransientChatTabKey(key));
-    if (sanitized.length > 0) {
-      next[workspaceId] = sanitized;
-    }
-  }
-  return next;
-}
-
-function sanitizeSessionIdArrayRecord(value: unknown): Record<string, string[]> {
-  if (!isStringArrayRecord(value)) {
-    return {};
-  }
-  return Object.fromEntries(
-    Object.entries(value).flatMap(([workspaceId, sessionIds]) => {
-      const sanitized = uniqueIds(sessionIds.filter((sessionId) =>
-        !isTransientClientSessionId(sessionId)
-      ));
-      return sanitized.length > 0 ? [[workspaceId, sanitized]] : [];
-    }),
-  );
-}
-
-function sanitizeLastViewedSessionByWorkspace(value: unknown): Record<string, string> {
-  if (!isStringRecord(value)) {
-    return {};
-  }
-  return Object.fromEntries(
-    Object.entries(value).filter(([, sessionId]) => !isTransientClientSessionId(sessionId)),
-  );
-}
-
-function sanitizeManualChatGroupsWithoutTransientSessions(
-  value: Record<string, ManualChatGroup[]>,
-): Record<string, ManualChatGroup[]> {
-  return Object.fromEntries(
-    Object.entries(value).flatMap(([workspaceId, groups]) => {
-      const nextGroups = groups.flatMap((group) => {
-        const sessionIds = uniqueIds(group.sessionIds.filter((sessionId) =>
-          !isTransientClientSessionId(sessionId)
-        ));
-        return sessionIds.length > 0 ? [{ ...group, sessionIds }] : [];
-      });
-      return nextGroups.length > 0 ? [[workspaceId, nextGroups]] : [];
-    }),
-  );
-}
-
-function isTransientChatTabKey(key: string): boolean {
-  const parsed = parseWorkspaceShellTabKey(key);
-  return parsed?.kind === "chat" && isTransientClientSessionId(parsed.sessionId);
-}
-
-function isTransientClientSessionId(sessionId: string): boolean {
-  return sessionId.startsWith("client-session:")
-    || sessionId.startsWith("pending-session:");
-}
-
-function sanitizeRightPanelWidths(value: unknown): Record<string, number> {
-  if (typeof value !== "object" || value === null) {
-    return {};
-  }
-
-  const next: Record<string, number> = {};
-  for (const [workspaceId, width] of Object.entries(value)) {
-    if (typeof width === "number" && Number.isFinite(width)) {
-      next[workspaceId] = clampRightPanelWidth(width);
-    }
-  }
-  return next;
 }
 
 function rightPanelStateUpdate(
@@ -782,6 +200,13 @@ export const useWorkspaceUiStore = create<WorkspaceUiState>((set, get) => ({
   shellActivationEpochByWorkspace: {},
   pendingChatActivationByWorkspace: {},
   urgentHighlightedChatSessionByWorkspace: {},
+
+  hydrate: (state) => {
+    set({
+      ...state,
+      _hydrated: true,
+    });
+  },
 
   archiveWorkspace: (id) => {
     const current = get().archivedWorkspaceIds;
@@ -844,7 +269,7 @@ export const useWorkspaceUiStore = create<WorkspaceUiState>((set, get) => ({
 
   setSidebarWidth: (value) => {
     set((state) => ({
-      sidebarWidth: clampSidebarWidth(resolveStateValue(value, state.sidebarWidth)),
+      sidebarWidth: clampWorkspaceSidebarWidth(resolveStateValue(value, state.sidebarWidth)),
     }));
   },
 
@@ -916,124 +341,7 @@ export const useWorkspaceUiStore = create<WorkspaceUiState>((set, get) => ({
     });
   },
 
-  setActiveShellTabKeyForWorkspace: (workspaceId, key) => {
-    get().writeShellIntent({ workspaceId, intent: key });
-  },
-
-  writeShellIntent: ({ workspaceId, intent }) => {
-    const hasCurrent = Object.prototype.hasOwnProperty.call(
-      get().activeShellTabKeyByWorkspace,
-      workspaceId,
-    );
-    const current = hasCurrent ? get().activeShellTabKeyByWorkspace[workspaceId] : null;
-    const previousEpoch = get().shellActivationEpochByWorkspace[workspaceId] ?? 0;
-    if (hasCurrent && current === intent) {
-      return {
-        changed: false,
-        previousIntent: current,
-        currentIntent: current,
-        epoch: previousEpoch,
-      };
-    }
-    const nextEpoch = previousEpoch + 1;
-    set({
-      activeShellTabKeyByWorkspace: {
-        ...get().activeShellTabKeyByWorkspace,
-        [workspaceId]: intent,
-      },
-      shellActivationEpochByWorkspace: {
-        ...get().shellActivationEpochByWorkspace,
-        [workspaceId]: nextEpoch,
-      },
-    });
-    return {
-      changed: true,
-      previousIntent: current,
-      currentIntent: intent,
-      epoch: nextEpoch,
-    };
-  },
-
-  replaceShellIntent: ({ workspaceId, expectedIntent, nextIntent, expectedEpoch }) => {
-    const previousIntent = get().activeShellTabKeyByWorkspace[workspaceId] ?? null;
-    const previousEpoch = get().shellActivationEpochByWorkspace[workspaceId] ?? 0;
-    if (
-      previousIntent !== expectedIntent
-      || (expectedEpoch !== undefined && previousEpoch !== expectedEpoch)
-    ) {
-      return {
-        changed: false,
-        replaced: false,
-        previousIntent,
-        currentIntent: previousIntent,
-        epoch: previousEpoch,
-      };
-    }
-    const result = get().writeShellIntent({ workspaceId, intent: nextIntent });
-    return { ...result, replaced: result.changed };
-  },
-
-  rollbackShellIntent: ({
-    workspaceId,
-    expectedIntent,
-    expectedEpoch,
-    expectedPendingAttemptId,
-    rollbackIntent,
-  }) => {
-    const previousIntent = get().activeShellTabKeyByWorkspace[workspaceId] ?? null;
-    const previousEpoch = get().shellActivationEpochByWorkspace[workspaceId] ?? 0;
-    const pending = get().pendingChatActivationByWorkspace[workspaceId] ?? null;
-    if (
-      previousIntent !== expectedIntent
-      || previousEpoch !== expectedEpoch
-      || (
-        expectedPendingAttemptId !== undefined
-        && pending?.attemptId !== expectedPendingAttemptId
-      )
-    ) {
-      return {
-        changed: false,
-        rolledBack: false,
-        previousIntent,
-        currentIntent: previousIntent,
-        epoch: previousEpoch,
-      };
-    }
-    const result = get().writeShellIntent({ workspaceId, intent: rollbackIntent });
-    return { ...result, rolledBack: result.changed };
-  },
-
-  setPendingChatActivation: ({ workspaceId, pending }) => {
-    set({
-      pendingChatActivationByWorkspace: {
-        ...get().pendingChatActivationByWorkspace,
-        [workspaceId]: pending,
-      },
-    });
-    return { set: true };
-  },
-
-  clearPendingChatActivation: ({ workspaceId, attemptId, bumpIfCurrent }) => {
-    const pending = get().pendingChatActivationByWorkspace[workspaceId] ?? null;
-    const epoch = get().shellActivationEpochByWorkspace[workspaceId] ?? 0;
-    if (!pending || pending.attemptId !== attemptId) {
-      return { cleared: false, bumped: false, epoch };
-    }
-    const nextEpoch = bumpIfCurrent ? epoch + 1 : epoch;
-    set({
-      pendingChatActivationByWorkspace: {
-        ...get().pendingChatActivationByWorkspace,
-        [workspaceId]: null,
-      },
-      shellActivationEpochByWorkspace: bumpIfCurrent
-        ? {
-          ...get().shellActivationEpochByWorkspace,
-          [workspaceId]: nextEpoch,
-        }
-        : get().shellActivationEpochByWorkspace,
-    });
-    return { cleared: true, bumped: bumpIfCurrent, epoch: nextEpoch };
-  },
+  ...createWorkspaceUiShellActions(set, get),
 
   setUrgentHighlightedChatSessionForWorkspace: (workspaceId, sessionId) => {
     set({
@@ -1054,40 +362,6 @@ export const useWorkspaceUiStore = create<WorkspaceUiState>((set, get) => ({
         ...get().urgentHighlightedChatSessionByWorkspace,
         [workspaceId]: null,
       },
-    });
-  },
-
-  setShellTabOrderForWorkspace: (workspaceId, order) => {
-    const hasCurrent = Object.prototype.hasOwnProperty.call(
-      get().shellTabOrderByWorkspace,
-      workspaceId,
-    );
-    const current = hasCurrent ? get().shellTabOrderByWorkspace[workspaceId] : [];
-    if (hasCurrent && sameStringArray(current, order)) {
-      return;
-    }
-    set({
-      shellTabOrderByWorkspace: {
-        ...get().shellTabOrderByWorkspace,
-        [workspaceId]: order,
-      },
-    });
-  },
-
-  resetWorkspaceShellTabs: (workspaceId) => {
-    const active = { ...get().activeShellTabKeyByWorkspace };
-    const order = { ...get().shellTabOrderByWorkspace };
-    const epoch = { ...get().shellActivationEpochByWorkspace };
-    const pending = { ...get().pendingChatActivationByWorkspace };
-    delete active[workspaceId];
-    delete order[workspaceId];
-    delete epoch[workspaceId];
-    delete pending[workspaceId];
-    set({
-      activeShellTabKeyByWorkspace: active,
-      shellTabOrderByWorkspace: order,
-      shellActivationEpochByWorkspace: epoch,
-      pendingChatActivationByWorkspace: pending,
     });
   },
 
@@ -1256,209 +530,8 @@ export const useWorkspaceUiStore = create<WorkspaceUiState>((set, get) => ({
     set({ finishSuggestionDismissalsByWorkspaceId: current });
   },
 
-  setVisibleChatSessionIdsForWorkspace: (workspaceId, sessionIds) => {
-    const nextSessionIds = uniqueIds(sessionIds);
-    set((state) => {
-      const hasCurrent = Object.prototype.hasOwnProperty.call(
-        state.visibleChatSessionIdsByWorkspace,
-        workspaceId,
-      );
-      const current = state.visibleChatSessionIdsByWorkspace[workspaceId] ?? [];
-      if (hasCurrent && sameStringArray(current, nextSessionIds)) {
-        return state;
-      }
-      return {
-        visibleChatSessionIdsByWorkspace: {
-          ...state.visibleChatSessionIdsByWorkspace,
-          [workspaceId]: nextSessionIds,
-        },
-      };
-    });
-  },
-
-  rememberHiddenChatSessionForWorkspace: (workspaceId, sessionId) => {
-    const current =
-      get().recentlyHiddenChatSessionIdsByWorkspace[workspaceId] ?? [];
-    set({
-      recentlyHiddenChatSessionIdsByWorkspace: {
-        ...get().recentlyHiddenChatSessionIdsByWorkspace,
-        [workspaceId]: rememberHiddenChatSessionId(current, sessionId),
-      },
-    });
-  },
-
-  clearHiddenChatSessionsForWorkspace: (workspaceId, sessionIds) => {
-    const current =
-      get().recentlyHiddenChatSessionIdsByWorkspace[workspaceId] ?? [];
-    const next = clearHiddenChatSessionIds(current, sessionIds);
-    if (next.length === current.length) {
-      return;
-    }
-    set({
-      recentlyHiddenChatSessionIdsByWorkspace: {
-        ...get().recentlyHiddenChatSessionIdsByWorkspace,
-        [workspaceId]: next,
-      },
-    });
-  },
-
-  toggleChatGroupCollapsedForWorkspace: (workspaceId, parentSessionId) => {
-    const current =
-      get().collapsedChatGroupsByWorkspace[workspaceId] ?? [];
-    const next = current.includes(parentSessionId)
-      ? current.filter((id) => id !== parentSessionId)
-      : uniqueIds([...current, parentSessionId]);
-    const collapsed = { ...get().collapsedChatGroupsByWorkspace };
-    if (next.length > 0) {
-      collapsed[workspaceId] = next;
-    } else {
-      delete collapsed[workspaceId];
-    }
-    set({ collapsedChatGroupsByWorkspace: collapsed });
-  },
-
-  clearChatGroupCollapsedForWorkspace: (workspaceId, parentSessionIds) => {
-    const current =
-      get().collapsedChatGroupsByWorkspace[workspaceId] ?? [];
-    if (current.length === 0 || parentSessionIds.length === 0) {
-      return;
-    }
-    const clearSet = new Set(parentSessionIds);
-    const next = current.filter((id) => !clearSet.has(id));
-    if (next.length === current.length) {
-      return;
-    }
-    const collapsed = { ...get().collapsedChatGroupsByWorkspace };
-    if (next.length > 0) {
-      collapsed[workspaceId] = next;
-    } else {
-      delete collapsed[workspaceId];
-    }
-    set({ collapsedChatGroupsByWorkspace: collapsed });
-  },
-
-  setManualChatGroupsForWorkspace: (workspaceId, groups) => {
-    const current = get().manualChatGroupsByWorkspace;
-    const nextGroupsByWorkspace = { ...current };
-    if (groups.length > 0) {
-      nextGroupsByWorkspace[workspaceId] = groups;
-    } else {
-      delete nextGroupsByWorkspace[workspaceId];
-    }
-    set({ manualChatGroupsByWorkspace: nextGroupsByWorkspace });
-  },
-
-  upsertManualChatGroupForWorkspace: (workspaceId, group) => {
-    const current = get().manualChatGroupsByWorkspace[workspaceId] ?? [];
-    const nextGroups = upsertManualChatGroup(current, group);
-    const nextGroupsByWorkspace = { ...get().manualChatGroupsByWorkspace };
-    if (nextGroups.length > 0) {
-      nextGroupsByWorkspace[workspaceId] = nextGroups;
-    } else {
-      delete nextGroupsByWorkspace[workspaceId];
-    }
-    set({ manualChatGroupsByWorkspace: nextGroupsByWorkspace });
-  },
-
-  updateManualChatGroupForWorkspace: (workspaceId, groupId, updates) => {
-    const current = get().manualChatGroupsByWorkspace[workspaceId] ?? [];
-    const nextGroups = updateManualChatGroup(current, groupId, updates);
-    const nextGroupsByWorkspace = { ...get().manualChatGroupsByWorkspace };
-    if (nextGroups.length > 0) {
-      nextGroupsByWorkspace[workspaceId] = nextGroups;
-    } else {
-      delete nextGroupsByWorkspace[workspaceId];
-    }
-    set({ manualChatGroupsByWorkspace: nextGroupsByWorkspace });
-  },
-
-  deleteManualChatGroupForWorkspace: (workspaceId, groupId) => {
-    const currentGroups = get().manualChatGroupsByWorkspace[workspaceId] ?? [];
-    const nextGroups = deleteManualChatGroup(currentGroups, groupId);
-    const nextGroupsByWorkspace = { ...get().manualChatGroupsByWorkspace };
-
-    if (nextGroups.length > 0) {
-      nextGroupsByWorkspace[workspaceId] = nextGroups;
-    } else {
-      delete nextGroupsByWorkspace[workspaceId];
-    }
-    set({ manualChatGroupsByWorkspace: nextGroupsByWorkspace });
-  },
-
-  removeSessionsFromManualChatGroupsForWorkspace: (workspaceId, sessionIds) => {
-    const currentGroups = get().manualChatGroupsByWorkspace[workspaceId] ?? [];
-    const nextGroups = removeSessionsFromManualChatGroups(currentGroups, sessionIds);
-    const nextGroupsByWorkspace = { ...get().manualChatGroupsByWorkspace };
-
-    if (nextGroups.length > 0) {
-      nextGroupsByWorkspace[workspaceId] = nextGroups;
-    } else {
-      delete nextGroupsByWorkspace[workspaceId];
-    }
-    set({ manualChatGroupsByWorkspace: nextGroupsByWorkspace });
-  },
-
-  clearWorkspaceChatTabState: (workspaceId) => {
-    const visible = { ...get().visibleChatSessionIdsByWorkspace };
-    const hidden = { ...get().recentlyHiddenChatSessionIdsByWorkspace };
-    const collapsed = { ...get().collapsedChatGroupsByWorkspace };
-    const manualGroups = { ...get().manualChatGroupsByWorkspace };
-    delete visible[workspaceId];
-    delete hidden[workspaceId];
-    delete collapsed[workspaceId];
-    delete manualGroups[workspaceId];
-    set({
-      visibleChatSessionIdsByWorkspace: visible,
-      recentlyHiddenChatSessionIdsByWorkspace: hidden,
-      collapsedChatGroupsByWorkspace: collapsed,
-      manualChatGroupsByWorkspace: manualGroups,
-    });
-  },
+  ...createWorkspaceUiChatTabActions(set, get),
 }));
-
-useWorkspaceUiStore.subscribe((state, prev) => {
-  const changedKeys = getChangedWorkspaceUiStateKeys(prev, state);
-  if (isDebugMeasurementEnabled()) {
-    if (changedKeys.length > 0) {
-      recordMeasurementDiagnostic({
-        category: "workspace_ui_store.write",
-        label: "top_level_keys",
-        keys: changedKeys,
-        count: changedKeys.length,
-      });
-    }
-  }
-
-  if (!state._hydrated) {
-    return;
-  }
-
-  if (
-    changedKeys.length > 0
-    && changedKeys.every(isNonPersistedWorkspaceUiStateKey)
-  ) {
-    return;
-  }
-
-  const currentSlice = selectPersistedSlice(state);
-  const previousSlice = selectPersistedSlice(prev);
-  if (JSON.stringify(currentSlice) !== JSON.stringify(previousSlice)) {
-    void persistValue(WORKSPACE_UI_KEY, currentSlice);
-  }
-});
-
-export async function bootstrapWorkspaceUi(): Promise<void> {
-  const { state, didMigrate } = await readAll();
-  useWorkspaceUiStore.setState({
-    ...state,
-    _hydrated: true,
-  });
-  if (didMigrate) {
-    // Force-persist so migrationVersion is saved even when the migration
-    // itself was a no-op (e.g. workspaceLastInteracted was already empty).
-    void persistValue(WORKSPACE_UI_KEY, selectPersistedSlice(useWorkspaceUiStore.getState()));
-  }
-}
 
 export function trackWorkspaceInteraction(workspaceId: string, timestamp: string) {
   useWorkspaceUiStore.getState().updateWorkspaceLastInteracted(workspaceId, timestamp);
@@ -1490,48 +563,4 @@ export function clearViewedSessionErrors(sessionIds: string[]) {
 
 export function ensureRepoGroupExpanded(repoKey: string) {
   useWorkspaceUiStore.getState().ensureRepoGroupExpanded(repoKey);
-}
-
-function getChangedWorkspaceUiStateKeys(
-  previous: WorkspaceUiState,
-  next: WorkspaceUiState,
-): string[] {
-  // Manual top-level allowlist for debug diagnostics. Keep this in sync when
-  // adding workspace UI state that should show up in store-write traces.
-  return [
-    "archivedWorkspaceIds",
-    "hiddenRepoRootIds",
-    "collapsedRepoGroups",
-    "showArchived",
-    "threadsCollapsed",
-    "sidebarOpen",
-    "sidebarWidth",
-    "rightPanelDurableByWorkspace",
-    "rightPanelMaterializedByWorkspace",
-    "activeShellTabKeyByWorkspace",
-    "shellTabOrderByWorkspace",
-    "shellActivationEpochByWorkspace",
-    "pendingChatActivationByWorkspace",
-    "urgentHighlightedChatSessionByWorkspace",
-    "workspaceTypes",
-    "lastViewedAt",
-    "lastViewedSessionByWorkspace",
-    "lastViewedSessionErrorAtBySession",
-    "workspaceLastInteracted",
-    "dismissedSetupFailures",
-    "finishSuggestionDismissalsByWorkspaceId",
-    "visibleChatSessionIdsByWorkspace",
-    "recentlyHiddenChatSessionIdsByWorkspace",
-    "collapsedChatGroupsByWorkspace",
-    "manualChatGroupsByWorkspace",
-  ].filter((key) => !Object.is(
-    previous[key as keyof WorkspaceUiState],
-    next[key as keyof WorkspaceUiState],
-  ));
-}
-
-function isNonPersistedWorkspaceUiStateKey(key: string): boolean {
-  return key === "pendingChatActivationByWorkspace"
-    || key === "shellActivationEpochByWorkspace"
-    || key === "urgentHighlightedChatSessionByWorkspace";
 }
