+++ scripts/max_lines_allowlist.txt
@@ -49,10 +49,8 @@ desktop/src/hooks/sessions/use-session-selection-actions.ts
 desktop/src/hooks/workspaces/tabs/use-workspace-header-tabs-view-model.ts
 desktop/src/hooks/workspaces/use-workspace-bootstrap-actions.ts
 desktop/src/lib/domain/chat/__fixtures__/playground.ts
-desktop/src/lib/domain/preferences/user-preferences.ts
 desktop/src/lib/infra/measurement/debug-measurement.ts
 desktop/src/stores/preferences/workspace-ui-store.test.ts
-desktop/src/stores/preferences/workspace-ui-store.ts
 server/tests/integration/test_cloud_api.py
 server/tests/integration/test_billing_accounting.py
 server/tests/integration/test_billing_api.py
